feat: automatically save progress
Automatically saves the progress of the user. If the saved progress file is present, the app reads from that file, if not, it reads from the source file.

Changed the main data variable and handling to keep the cards as a DataFrame instead of a dictionary, so the read-write process is more easy to follow and implement.

Fixed some typos.